backend/requirements.txt
# GameSpot Booking System - Python Backend Requirements

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# MySQL Database Connector
mysql-connector-python==8.2.0

# Security
Werkzeug==3.0.1
bcrypt==4.1.2
PyJWT==2.8.0
pycryptodome>=3.19.0

# OAuth
google-auth==2.25.2
google-auth-oauthlib==1.2.0
requests==2.31.0

# SMS
twilio==8.11.0
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Date/Time Handling
python-dateutil==2.8.2

# 🚀 AI & LLM Integration
google-generativeai==0.3.2

# 🎤 VOICE TTS (Text-to-Speech) - Multiple Options
# Install based on preference:

# Option 1: Edge TTS (FREE, Best balance) - RECOMMENDED
edge-tts==6.1.9

# Option 2: gTTS (FREE, Basic fallback) - Always install as backup
gTTS==2.5.0

# Option 3: Google Cloud TTS (Excellent quality, free tier)
# google-cloud-texttospeech==2.14.1

# Option 4: OpenAI TTS (Best quality, requires API key)
# openai==1.10.0

# Option 5: ElevenLabs (Ultra realistic, requires API key)
# elevenlabs==0.2.27

# Option 6: Coqui TTS (Local, FREE, requires download)
# TTS==0.22.0

# AI Integration - Google Gemini (FREE)
google-generativeai==0.3.2


gunicorn==21.2.0
